Fix: Parse as text for non-binary payload (#42)
When using tools such as Postman or curl to send events to AEM, payload
is sent as normal text/json and not as binary. So on the receiver side
(plugin) it has to be treated as such. See
https://community.solace.com/t/weird-chars-at-start-of-text-payload/1856.

When using getBinaryAttachment() on a string payload, you will get the
following:
```
Sending:
{ "foo":"bar" }

Receiving:
" N{ \"foo\":\"bar\" }"
```

With 2 invalid characters in the beginning, which are payload-length
identifiers from the Structured Data Type (SDT) formatted message.